SendMoney Page UI Inhancementand Registration Page - 20/02/2025
