+++ app/src/main/java/com/example/accbankandroid/Registration.kt
@@ -1,0 +1,201 @@
+package com.example.accbankandroid
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.shape.RoundedCornerShape
+import androidx.compose.foundation.verticalScroll
+import androidx.compose.material3.Button
+import androidx.compose.material3.ButtonDefaults
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.Icon
+import androidx.compose.material3.IconButton
+import androidx.compose.material3.OutlinedTextField
+import androidx.compose.material3.Text
+import androidx.compose.material3.TextFieldDefaults
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import com.example.accbankandroid.ui.theme.getGradientBrush
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.text.font.FontFamily
+import androidx.compose.ui.text.input.PasswordVisualTransformation
+import androidx.compose.ui.text.input.VisualTransformation
+import androidx.compose.ui.text.style.TextAlign
+import androidx.compose.ui.tooling.preview.Preview
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+
+@OptIn(ExperimentalMaterial3Api::class)
+@Preview
+@Composable
+fun registration()
+{
+    var email by remember { mutableStateOf("") }
+    var password by remember { mutableStateOf("") }
+    var passwordVisible by remember { mutableStateOf(false) }
+    var confirmPassword by remember { mutableStateOf("") }
+    var confirmPasswordVisible by remember { mutableStateOf(false) }
+    var errorMessage by remember { mutableStateOf("") }
+
+    // ✅ Remember scroll state for vertical scrolling
+    val scrollState = rememberScrollState()
+
+
+    Box(
+        modifier = Modifier
+            .fillMaxSize()
+            .background(getGradientBrush()) // Apply new gradient
+            .padding(16.dp),
+        contentAlignment = Alignment.Center
+    ){
+
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .verticalScroll(scrollState) // ✅ Enables scrolling
+                .padding(20.dp),
+            horizontalAlignment = Alignment.CenterHorizontally
+        ) {
+            //  Glassmorphic Email Field
+            OutlinedTextField(
+                value = email,
+                onValueChange = { email = it },
+                label = { Text("User ID/username") },
+                singleLine = true,
+                shape = RoundedCornerShape(24.dp),
+                modifier = Modifier
+                    .fillMaxWidth(),
+                trailingIcon = {
+                    IconButton(onClick = { /* Action to remember username */ }) {
+                        Icon(painter = painterResource(id = android.R.drawable.ic_dialog_info), contentDescription = "Info", tint = Color.White)
+                    }
+                },
+                colors = TextFieldDefaults.outlinedTextFieldColors(
+                    focusedTextColor = Color.White,
+                    unfocusedTextColor = Color.White,
+                    unfocusedLabelColor = Color.White,
+                    focusedLabelColor = Color.White,
+                    focusedBorderColor = Color.White.copy(alpha = 0.9f),
+                    unfocusedBorderColor = Color.White.copy(alpha = 0.5f),
+                    containerColor = Color.White.copy(alpha = 0.1f)
+                )
+            )
+
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            // Password Field with Eye Icon
+            OutlinedTextField(
+                value = password,
+                onValueChange = { password = it },
+                label = { Text("Password") },
+                singleLine = true,
+                shape = RoundedCornerShape(24.dp),
+                visualTransformation = if (passwordVisible) VisualTransformation.None else PasswordVisualTransformation(),
+                modifier = Modifier.fillMaxWidth(),
+                trailingIcon = {
+                    IconButton(onClick = { passwordVisible = !passwordVisible }) {
+                        Icon(painter = painterResource(id = android.R.drawable.ic_menu_view), contentDescription = "Toggle Password Visibility", tint = Color.White)
+                    }
+                },
+                colors = TextFieldDefaults.outlinedTextFieldColors(
+                    focusedTextColor = Color.White,
+                    unfocusedTextColor = Color.White,
+                    unfocusedLabelColor = Color.White,
+                    focusedLabelColor = Color.White,
+                    focusedBorderColor = Color.White.copy(alpha = 0.9f),
+                    unfocusedBorderColor = Color.White.copy(alpha = 0.5f),
+                    containerColor = Color.White.copy(alpha = 0.1f)
+                )
+            )
+
+            Spacer(modifier = Modifier.height(16.dp))
+
+            // Confirm Password Field with Eye Icon
+            OutlinedTextField(
+                value = confirmPassword,
+                onValueChange = { confirmPassword = it },
+                label = { Text("Confirm Password") },
+                singleLine = true,
+                shape = RoundedCornerShape(24.dp),
+                visualTransformation = if (confirmPasswordVisible) VisualTransformation.None else PasswordVisualTransformation(),
+                modifier = Modifier.fillMaxWidth(),
+                trailingIcon = {
+                    IconButton(onClick = { confirmPasswordVisible = !confirmPasswordVisible }) {
+                        Icon(painter = painterResource(id = android.R.drawable.ic_menu_view), contentDescription = "Toggle Password Visibility", tint = Color.White)
+                    }
+                },
+                colors = TextFieldDefaults.outlinedTextFieldColors(
+                    focusedTextColor = Color.White,
+                    unfocusedTextColor = Color.White,
+                    unfocusedLabelColor = Color.White,
+                    focusedLabelColor = Color.White,
+                    focusedBorderColor = Color.White.copy(alpha = 0.9f),
+                    unfocusedBorderColor = Color.White.copy(alpha = 0.5f),
+                    containerColor = Color.White.copy(alpha = 0.1f)
+                )
+            )
+            if (errorMessage.isNotEmpty()) {
+                Column(
+                    modifier = Modifier.fillMaxWidth(),
+                    horizontalAlignment = Alignment.CenterHorizontally
+                ) {
+                    Text(
+                        text = errorMessage,
+                        color = Color.Red,
+                        fontSize = 16.sp,
+                        fontFamily = FontFamily.SansSerif,
+                        textAlign = TextAlign.Center,
+                        modifier = Modifier
+                            .padding(top = 4.dp)
+                            .fillMaxWidth()
+                    )
+                }
+            }
+
+            Spacer(modifier = Modifier.height(32.dp))
+
+            Button(
+                onClick = {
+                    if (password.isEmpty() || confirmPassword.isEmpty()) {
+                        errorMessage = "password cannot be empty"
+                     } else if (password != confirmPassword) {
+                         errorMessage = "password and Confirm password do not match"
+                    } else {
+                        errorMessage = "" // Clear error when valid
+                        // Proceed with submission
+                    }
+                          },
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(50.dp),
+                shape = RoundedCornerShape(50.dp), // Fully rounded
+                elevation = ButtonDefaults.buttonElevation(defaultElevation = 8.dp),
+                colors = ButtonDefaults.buttonColors(containerColor = Color.White)
+            ) {
+                Text(text = "Register", fontSize = 18.sp, color = Color.Black)
+            }
+            Spacer(modifier = Modifier.height(100.dp))
+
+            /*Box(
+                modifier = Modifier
+                    .width(300.dp) // Underline length
+                    .height(2.dp) // Thickness
+                    .background(Color.White.copy(alpha = 0.5f))
+            )*/
+        }
+    }
+}
