+++ app/src/main/java/com/example/accbankandroid/SendMoney.kt
@@ -185,27 +185,6 @@ fun SendMoney(modifier: Modifier = Modifier){
                 }
             }
         }
-        when (selectedTabIndex) {
-            0 -> NewTabContent() // Show this when "New" is selected
-            1 -> HistoryTabContent() // Show this when "History" is selected
-        }
-    }
-    }
-
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun NewTabContent() {
-
-    var amount by remember { mutableStateOf("$503.66") }
-    var notifyMethod by remember { mutableStateOf("Email & SMS") }
-    var securityquestion by remember { mutableStateOf("Question") }
-    var securityanswer by remember { mutableStateOf("***********") }
-    var confirmsecurityanswer by remember { mutableStateOf("***********") }
-    var message by remember { mutableStateOf("") } // State to store the input text
-    val minLength = 250 // Minimum length for the message
-
-    val scrollState = rememberScrollState()//scroll
-
         Spacer(modifier = Modifier.padding(3.dp))
         Text(
             text = ("VIA INTERAC E-TRANSFER"),
@@ -216,352 +195,11 @@ fun NewTabContent() {
             textAlign = TextAlign.Left,
             modifier = Modifier.fillMaxWidth().padding(start = 21.dp) // Ensures alignment works
         )
-        Spacer(modifier = Modifier.height(8.dp))
-        Column(
-            modifier = Modifier
-                .fillMaxSize()
-                .verticalScroll(scrollState) // Makes content scrollable
-                .padding(10.dp) // Adds outer padding
-        ) {
-            Card(
-                modifier = Modifier
-                    .fillMaxWidth()
-                    .padding(10.dp),  // Adds padding around the card
-                shape = RoundedCornerShape(22.dp),  // Rounded corners
-                colors = CardDefaults.cardColors(containerColor = Color.White),  // White background
-                elevation = CardDefaults.cardElevation(defaultElevation = 7.dp),  // Adds a slight shadow
-                border = BorderStroke(1.dp, Color.LightGray) , // Adds a subtle border
-            ) {
-                Column(
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .padding(12.dp)
-                ) {
-                    Spacer(modifier = Modifier.height(15.dp))
-                    Column(
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        Text(
-                            text = "FROM",
-                            fontSize = 14.sp,
-                            fontWeight = FontWeight.Medium,
-                            color = Color.Gray,
-                            fontFamily = FontFamily.SansSerif,
-                            letterSpacing = 2.sp  // Adjust this value as needed for spacing
-                        )
-                        Spacer(modifier = Modifier.height(5.dp))
-                        Row(
-                            modifier = Modifier.fillMaxWidth(),
-                            horizontalArrangement = Arrangement.SpaceBetween
-                        ) {
-                            Text("Chequing", fontSize = 18.sp, fontWeight = FontWeight.Bold,fontFamily = FontFamily.SansSerif)
-                            Text(
-                                "$1245.45",
-                                fontSize = 16.sp,
-                                color = Color(0xFF2F8F59),
-                                fontWeight = FontWeight.Bold,
-                                fontFamily = FontFamily.SansSerif,
-                                letterSpacing = 1.sp
-                            )
-                        }
-                        Text("******1234", fontSize = 16.sp, color = Color.Gray,fontFamily = FontFamily.SansSerif)
-                    }
-                    Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                    Divider(color = Color.Gray, thickness = 1.dp)
-                    Column(
-                        modifier = Modifier.fillMaxWidth()
-                    ) {
-                        Spacer(modifier = Modifier.height(10.dp))
-                        Text(
-                            text = "SEND TO ",
-                            fontSize = 14.sp,
-                            fontWeight = FontWeight.Medium,
-                            color = Color.Gray,
-                            fontFamily = FontFamily.SansSerif,
-                            letterSpacing = 2.sp  // Adjust this value as needed for spacing
-                        )
-                        Spacer(modifier = Modifier.height(5.dp))
-                        Text("SAM Peter", fontSize = 18.sp, fontWeight = FontWeight.Bold,fontFamily = FontFamily.SansSerif)
-                        Text("sam_test@gmail.com", fontSize = 16.sp, color = Color.Gray,fontFamily = FontFamily.SansSerif)
-                        Text("(416) 555-1234", fontSize = 16.sp, color = Color.Gray,fontFamily = FontFamily.SansSerif)
-                    }
-                    Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                    Divider(color = Color.Gray, thickness = 1.dp)
-                    //Spacer(modifier = Modifier.height(10.dp))
-
-                }
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    TextField(
-                        value = amount,
-                        onValueChange = {
-                            amount = it.filter { char -> char.isDigit() || char == '.' } // Allow only numbers and decimals
-                        },
-                        label = { Text("HOW MUCH",fontFamily = FontFamily.SansSerif) },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .background(Color.Transparent),
-                        textStyle = TextStyle( // Apply bold styling
-                            fontWeight = FontWeight.Bold,
-                            fontSize = 18.sp,
-                            fontFamily = FontFamily.SansSerif,
-                            color = Color.Black // Ensure it's readable
-                        ),
-                        keyboardOptions = KeyboardOptions(
-                            keyboardType = KeyboardType.Number // Opens numeric keyboard
-                        ),
-                        colors = TextFieldDefaults.textFieldColors(
-                            containerColor = Color.White, // Set background to pure white
-                            focusedIndicatorColor = Color.Transparent, // Remove underline
-                            unfocusedIndicatorColor = Color.Transparent
-                        )
-                    )
-                }
-                Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                Divider(
-                    color = Color.Gray,
-                    thickness = 1.dp,
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                )
-
-                NotifyRecipientDropdown()
-
-                Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                Divider(
-                    color = Color.Gray,
-                    thickness = 1.dp,
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                )
-                Spacer(modifier = Modifier.height(10.dp))
-
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    TextField(
-                        value = securityquestion,
-                        onValueChange = { securityquestion = it },
-                        label = { Text("SECURITY QUESTION",fontFamily = FontFamily.SansSerif) },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .background(Color.Transparent),
-                        textStyle = TextStyle( // Apply bold styling
-                            fontWeight = FontWeight.Bold,
-                            fontSize = 18.sp,
-                            fontFamily = FontFamily.SansSerif,
-                            color = Color.Black // Ensure it's readable
-                        ),
-                        colors = TextFieldDefaults.textFieldColors(
-                            containerColor = Color.White, // Set background to pure white
-                            focusedIndicatorColor = Color.Transparent, // Remove underline
-                            unfocusedIndicatorColor = Color.Transparent
-                        )
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                Divider(
-                    color = Color.Gray,
-                    thickness = 1.dp,
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                )
-                Spacer(modifier = Modifier.height(10.dp))
-
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    TextField(
-                        value = securityanswer,
-                        onValueChange = { securityanswer = it },
-                        visualTransformation = PasswordVisualTransformation(),
-                        label = { Text("SECURITY ANSWER",fontFamily = FontFamily.SansSerif) },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .background(Color.Transparent),
-                        textStyle = TextStyle( // Apply bold styling
-                            fontWeight = FontWeight.Bold,
-                            fontSize = 18.sp,
-                            fontFamily = FontFamily.SansSerif,
-                            color = Color.Black // Ensure it's readable
-                        ),
-                        colors = TextFieldDefaults.textFieldColors(
-                            containerColor = Color.White, // Set background to pure white
-                            focusedIndicatorColor = Color.Transparent, // Remove underline
-                            unfocusedIndicatorColor = Color.Transparent
-                        )
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                Divider(
-                    color = Color.Gray,
-                    thickness = 1.dp,
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                )
-                Spacer(modifier = Modifier.height(10.dp))
-
-                Column(
-                    modifier = Modifier.fillMaxWidth(),
-                ) {
-                    TextField(
-                        value = confirmsecurityanswer,
-                        onValueChange = { confirmsecurityanswer = it },
-                        visualTransformation = PasswordVisualTransformation(),
-                        label = { Text("CONFIRM SECURITY ANSWER",fontFamily = FontFamily.SansSerif) },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .background(Color.Transparent),
-                        textStyle = TextStyle( // Apply bold styling
-                            fontWeight = FontWeight.Bold,
-                            fontSize = 18.sp,
-                            fontFamily = FontFamily.SansSerif,
-                            color = Color.Black // Ensure it's readable
-                        ),
-                        colors = TextFieldDefaults.textFieldColors(
-                            containerColor = Color.White, // Set background to pure white
-                            focusedIndicatorColor = Color.Transparent, // Remove underline
-                            unfocusedIndicatorColor = Color.Transparent
-                        )
-                    )
-                }
-
-                Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                Divider(
-                    color = Color.Gray,
-                    thickness = 1.dp,
-                    modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                )
-                Spacer(modifier = Modifier.height(10.dp))
-
-                Column(
-                    modifier = Modifier.fillMaxWidth()
-                ) {
-                    TextField(
-                        value = message,
-                        onValueChange = { message = it },
-                        label = { Text("MESSAGE") },
-                        modifier = Modifier
-                            .fillMaxWidth()
-                            .height(80.dp),
-                        textStyle = TextStyle(fontSize = 16.sp),
-                        colors = TextFieldDefaults.textFieldColors(
-                            containerColor = Color.White,
-                            focusedIndicatorColor = Color.Transparent,
-                            unfocusedIndicatorColor = Color.Transparent
-                        ),
-                        placeholder = { Text("Optional (Maximum Characters 250)", color = Color.Gray) }
-                    )
-                    Spacer(modifier = Modifier.height(10.dp)) // Space before the line
-                    Divider(
-                        color = Color.Gray,
-                        thickness = 1.dp,
-                        modifier = Modifier.padding(horizontal = 16.dp, vertical = 8.dp)
-                    )
-                    Spacer(modifier = Modifier.height(20.dp))
-                }
-
-            }
-            Spacer(modifier = Modifier.height(10.dp))
-            Button(
-                onClick = { /* Handle click */ },
-                shape = RoundedCornerShape(20.dp), // Rounded corners
-                border = BorderStroke(1.dp, Color.White), // White border
-                colors = ButtonDefaults.buttonColors(
-                    containerColor = Color.Transparent // Button background is transparent
-                ),
-                modifier = Modifier
-                    .height(70.dp) // Button height
-                    .width(170.dp) // Button width
-                    .align(Alignment.CenterHorizontally) // Center horizontally
-            ) {
-                Box(
-                    modifier = Modifier
-                        .fillMaxSize() // Ensure gradient covers entire button
-                        .background(Brush.verticalGradient(listOf(Color.Cyan, Color.Blue)), shape = RoundedCornerShape(20.dp)),
-                    contentAlignment = Alignment.Center // Center text
-                ) {
-                    Text("✔ Send", color = Color.White, fontWeight = FontWeight.Bold, fontSize = 18.sp,fontFamily = FontFamily.SansSerif)
-                }
-            }
-
+        when (selectedTabIndex) {
+            0 -> NewTabContent() // Show this when "New" is selected
+            1 -> HistoryTabContent() // Show this when "History" is selected
         }
 
-}
-@OptIn(ExperimentalMaterial3Api::class)
-@Composable
-fun NotifyRecipientDropdown() {
-    var expanded by remember { mutableStateOf(false) }
-    var selectedOption by remember { mutableStateOf("Email & SMS") }
-
-    val options = listOf("Email", "SMS", "Email & SMS")
-
-    ExposedDropdownMenuBox(
-        expanded = expanded,
-        onExpandedChange = { expanded = it }
-    ) {
-        TextField(
-            value = selectedOption,
-            onValueChange = {},
-            readOnly = true,
-            label = { Text("Notify Recipient By",fontFamily = FontFamily.SansSerif) },
-            modifier = Modifier
-                .fillMaxWidth()
-                .background(Color.Transparent)
-                .menuAnchor(),
-            textStyle = TextStyle(
-                fontWeight = FontWeight.Bold,
-                fontSize = 18.sp,
-                color = Color.Black,
-                fontFamily = FontFamily.SansSerif
-            ),
-            trailingIcon = {
-                Icon(
-                    imageVector = if (expanded) Icons.Default.KeyboardArrowUp else Icons.Default.KeyboardArrowDown,
-                    contentDescription = null
-                )
-            },
-            colors = TextFieldDefaults.textFieldColors(
-                containerColor = Color.White,
-                focusedIndicatorColor = Color.Transparent,
-                unfocusedIndicatorColor = Color.Transparent
-            )
-        )
-
-        ExposedDropdownMenu(
-            expanded = expanded,
-            onDismissRequest = { expanded = false },
-            modifier = Modifier.background(Color.White) // White dropdown background
-        ) {
-            options.forEach { option ->
-                DropdownMenuItem(
-                    text = {
-                        Row(verticalAlignment = Alignment.CenterVertically) {
-                            Icon(
-                                imageVector = if (selectedOption == option) Icons.Default.Check else Icons.Default.MailOutline,
-                                contentDescription = "Option Icon",
-                                tint = if (selectedOption == option) Color.Blue else Color.Gray,
-                                modifier = Modifier.size(20.dp)
-                            )
-                            Spacer(modifier = Modifier.width(8.dp))
-                            Text(
-                                text = option,
-                                fontSize = 16.sp,
-                                fontWeight = FontWeight.Bold,
-                                color = if (selectedOption == option) Color.Blue else Color.Black
-                            )
-                        }
-                    },
-                    onClick = {
-                        selectedOption = option
-                        expanded = false
-                    },
-                    modifier = Modifier
-                        .fillMaxWidth()
-                        .background(if (selectedOption == option) Color(0xFFE0F7FA) else Color.White)
-
-                )
-            }
-        }
     }
 }
